# Optimize interop, use NET5 to run tests