* use .NET 5 for running tests, benchmarks and REPL
* cache found property accessors for CLR interop
* cache created TypeReference properties instead of always creating them again
* fix bound function arguments creation
* fix issue found with .NET 5 that params arguments version of Console.WriteLine was chosen even for "WriteLine('text')", now caching methods and prioritizing them so params will be considered last
* fix object indexers to be checked before falling back to object properties when doing reflection
* fast paths for set and get, without need to create intermediate property descriptors

fixes #802
fixes #804
fixes #806