@lahma 2020-11-15
## Jint.Benchmark.InteropBenchmark

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |DelegatesCanBeSet|7.789 μs|0.1553 μs|1.5469|0.0000|0.0000|6497 B|
| **New** |	| **7.877 μs (+1%)** | **0.1511 μs** | **1.5469 (0%)** | **0.0000** | **0.0000** | **6497 B (0%)** |
| Old |ExtraParametersAreIgnored|11.188 μs|0.1718 μs|2.1719|0.0000|0.0000|9130 B|
| **New** |	| **11.192 μs (0%)** | **0.2218 μs** | **2.1719 (0%)** | **0.0000** | **0.0000** | **9130 B (0%)** |
| Old |GetObjectProperties|6.434 μs|0.0444 μs|1.4297|0.0000|0.0000|5985 B|
| **New** |	| **6.488 μs (+1%)** | **0.1018 μs** | **1.4297 (0%)** | **0.0000** | **0.0000** | **5985 B (0%)** |
| Old |InvokeObjectMethods|7.289 μs|0.1454 μs|1.6719|0.0000|0.0000|7001 B|
| **New** |	| **7.535 μs (+3%)** | **0.1438 μs** | **1.6719 (0%)** | **0.0000** | **0.0000** | **7001 B (0%)** |
| Old |SetObjectProperties|9.870 μs|0.0834 μs|1.9375|0.0000|0.0000|8153 B|
| **New** |	| **9.883 μs (0%)** | **0.1415 μs** | **1.9375 (0%)** | **0.0000** | **0.0000** | **8153 B (0%)** |
| Old |GetIndexUsingStringKey|16.537 μs|0.1249 μs|2.9688|0.0000|0.0000|12434 B|
| **New** |	| **17.435 μs (+5%)** | **0.2281 μs** | **2.9688 (0%)** | **0.0000** | **0.0000** | **12434 B (0%)** |
| Old |GenericMethods|18.336 μs|0.1679 μs|3.0938|0.0000|0.0000|13042 B|
| **New** |	| **17.628 μs (-4%)** | **0.2524 μs** | **3.0938 (0%)** | **0.0000** | **0.0000** | **12962 B (-1%)** |
| Old |MultiGenericTypes|42.357 μs|0.6102 μs|6.1667|0.0000|0.0000|26027 B|
| **New** |	| **35.533 μs (-16%)** | **0.2872 μs** | **6.0000 (-3%)** | **0.0000** | **0.0000** | **25147 B (-3%)** |
| Old |IndexOnList|12.201 μs|0.1819 μs|2.1563|0.0000|0.0000|9040 B|
| **New** |	| **12.631 μs (+4%)** | **0.1958 μs** | **2.1563 (0%)** | **0.0000** | **0.0000** | **9040 B (0%)** |
| Old |Trim|16.992 μs|0.1831 μs|3.1250|0.0000|0.0000|13162 B|
| **New** |	| **17.660 μs (+4%)** | **0.0389 μs** | **3.1250 (0%)** | **0.0000** | **0.0000** | **13162 B (0%)** |
| Old |DelegateAsFunction|7.545 μs|0.0591 μs|1.5156|0.0000|0.0000|6385 B|
| **New** |	| **7.867 μs (+4%)** | **0.1461 μs** | **1.5156 (0%)** | **0.0000** | **0.0000** | **6385 B (0%)** |
| Old |ConvertArrayToArrayInstance|16.371 μs|0.3261 μs|3.2188|0.0000|0.0000|13576 B|
| **New** |	| **16.377 μs (0%)** | **0.1780 μs** | **3.2188 (0%)** | **0.0000** | **0.0000** | **13576 B (0%)** |
| Old |ConvertListsToArrayInstance|19.281 μs|0.2483 μs|3.6250|0.0000|0.0000|15256 B|
| **New** |	| **19.200 μs (0%)** | **0.2107 μs** | **3.6250 (0%)** | **0.0000** | **0.0000** | **15256 B (0%)** |
| Old |ConvertArrayInstanceToArray|3.751 μs|0.0730 μs|1.1914|0.0000|0.0000|4992 B|
| **New** |	| **3.745 μs (0%)** | **0.0733 μs** | **1.1914 (0%)** | **0.0000** | **0.0000** | **4992 B (0%)** |
| Old |ResolvingConsoleWriteLine|15.758 μs|0.0524 μs|2.5625|0.0000|0.0000|10776 B|
| **New** |	| **6.328 μs (-60%)** | **0.1222 μs** | **1.7813 (-30%)** | **0.0000** | **0.0000** | **7472 B (-31%)** |
| Old |LoopWithNativeEnumerator|7.071 μs|0.0661 μs|1.5078|0.0000|0.0000|6329 B|
| **New** |	| **7.233 μs (+2%)** | **0.1352 μs** | **1.5078 (0%)** | **0.0000** | **0.0000** | **6329 B (0%)** |